cpp 02

+++ cpp_modules/module02/ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: pablo <pablo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 20:50:13 by pablo             #+#    #+#             */
-/*   Updated: 2021/11/16 18:48:59 by pablo            ###   ########.fr       */
+/*   Updated: 2022/01/09 23:52:49 by pablo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int main()
 	Fixed const c(42.42f);
 	Fixed const d(b);
 	a = Fixed(1234.4321f);
-	std::cout << "a is " << a << std::endl;
+	std::cout << "a is " << a.toFloat() << std::endl;
 	std::cout << "b is " << b << std::endl;
 	std::cout << "c is " << c << std::endl;
 	std::cout << "d is " << d << std::endl;
